Update get_ebooks_function.py
Added code to create a new branch and submit a pull request for Version 1 process report B. I tested (textbook_search_app_CH) and debugged parts of the code in VS. The app shows on the browser, and the links work. I looked through and noted the code edited by Professor Harding and made comments.